feat: add endpoits depositar, sacar, consultar saldo e extrato

+++ src/main/java/com/aplication/bankapi/dto/conta/SaldoResponse.java
@@ -1,8 +1,10 @@
 package com.aplication.bankapi.dto.conta;
 
+import java.math.BigDecimal;
+
 public record SaldoResponse(
         Long contaId,
-        double saldo
+        BigDecimal saldo
 ) {
     
 }
