+++ src/main/java/com/aplication/bankapi/exception/SaldoInsuficienteException.java
@@ -1,7 +1,9 @@
 package com.aplication.bankapi.exception;
 
+import java.math.BigDecimal;
+
 public class SaldoInsuficienteException extends RuntimeException {
-    public SaldoInsuficienteException(Long contaId, double saldoDisponivel) {
+    public SaldoInsuficienteException(Long contaId, BigDecimal saldoDisponivel) {
        super("Saldo insuficiente na conta " + contaId + ". Saldo disponível: " + saldoDisponivel);
     }
